+++ Assets/Core/Game/Domains/GamePlay/Shared/MatchData/Models/MatchPlayerModel.cs
@@ -5,11 +5,11 @@ namespace Core.Game.Domains.GamePlay.Shared
 {
     public class MatchPlayerModel
     {
-        public int PlayerId;
+        public ushort PlayerId;
         public string PlayerName;
         public PlayerSpaceshipStateS2C Spaceship;
 
-        public MatchPlayerModel(int playerId, string playerName, PlayerSpaceshipStateS2C spaceship)
+        public MatchPlayerModel(ushort playerId, string playerName, PlayerSpaceshipStateS2C spaceship)
         {
             PlayerId = playerId;
             PlayerName = playerName;
