+++ Assets/Core/Game/Domains/GamePlay/Shared/S2CModels/PlayerSpaceshipStateS2C.cs
@@ -8,7 +8,7 @@ namespace Core.Game.Domains.GamePlay.Shared.S2CModels
         public PlayerShootStateS2C Shoot;
         public PlayerHealthS2C Health;
 
-        public PlayerSpaceshipStateS2C(PlayerTransformStateS2C transform, float shootCooldown, int health)
+        public PlayerSpaceshipStateS2C(PlayerTransformStateS2C transform, float shootCooldown, ushort health)
         {
             Transform = transform;
             Shoot = new PlayerShootStateS2C(shootCooldown);
