+++ Assets/Core/Game/Domains/GamePlay/Simulation/Scripts/MatchModel/IMatchNetEventsDataService.cs
@@ -15,7 +15,7 @@ namespace Core.Game.Domains.GamePlay.Simulation.Scripts.MatchModel
         void StartSavingPlayerEvents(ushort playerId);
         void StopSavingPlayerEvents(ushort playerId);
         void AddBulletSpawnNetEvent(int onTick, ushort bulletId, ushort belongToPlayerId, Vector2 position, float bulletRadius);
-        void AddPlayerTakeDamageNetEvent(int onTick, ushort damagedPlayerId, int playerHealth, int hitDamage, bool isAlive);
+        void AddPlayerTakeDamageNetEvent(int onTick, ushort damagedPlayerId, ushort playerHealth, ushort hitDamage, bool isAlive);
         void AddBulletDestroyedNetEvent(int onTick, ushort bulletId, Vector2 position);
         void AddPlayerJoinAcceptedEvent(int onTick, PlayerStateS2C playerState, SimulationStateS2C simulationState);
         void RemoveAllEventsOlderThanTick(ushort playerId, int tick);
