+++ Assets/Core/Game/Domains/GamePlay/Simulation/Scripts/Physics/IPhysicsSimulator.cs
@@ -10,7 +10,7 @@ namespace Core.Game.Domains.GamePlay.Simulation.Scripts.Physics
         void InitEntryPoint();
         void Step(float deltaTime, int velocityIterations, int positionIterations);
         void SetPlayerVelocity(ushort playerId, Vector2 velocity);
-        void AddWall(int id, Vector2[] points);
+        void AddWall(ushort id, Vector2[] points);
         void AddPlayer(ushort id, ushort teamId, Vector2 position, Vector2 velocity, float radius);
         public IReadOnlyList<PhysicsCollisionEvent> GetCachedCollisions();
         public void ClearCachedCollisions();
